+++ smawg/tests/test_misc.py
@@ -2,10 +2,11 @@
 
 import json
 import unittest
+from typing import Any
 
-from jsonschema.exceptions import ValidationError
+from pydantic import ValidationError
 
-from smawg import Ability, Combo, InvalidAssets, Race, validate
+from smawg import Ability, Combo, Race, validate
 from smawg._metadata import ASSETS_DIR
 from smawg.tests.common import TINY_ASSETS
 
@@ -18,9 +19,8 @@ class TestAssets(unittest.TestCase):
         for file_name in ASSETS_DIR.glob("*.json"):
             with open(file_name) as assets_file:
                 assets = json.load(assets_file)
-            # Raises an error and fails the test
-            # if `assets_json` is invalid or contains undocumented keys.
-            validate(assets, strict=True)
+            # Raises an error and fails the test if `assets` are invalid.
+            validate(assets)
 
 
 class TestCombo(unittest.TestCase):
@@ -39,27 +39,27 @@ class TestValidate(unittest.TestCase):
 
     def test_invalid_assets(self) -> None:
         """Check if `validate()` raises an error on invalid assets."""
-        invalid_fields = [
+        invalid_fields: list[tuple[str, Any]] = [
             # Bad structure of nested objects:
-            (ValidationError, "races", [{"not a": "race"}]),
-            (ValidationError, "abilities", [{"not a": "ability"}]),
-            (ValidationError, "map", {"not a": "map"}),
+            ("races", [{"not a": "race"}]),
+            ("abilities", [{"not a": "ability"}]),
+            ("map", {"not a": "map"}),
             # Borders between non-existring tiles:
-            (ValidationError, "map", {
+            ("map", {
                 "tiles": [],
                 "tile_borders": [[-2, -1]]
             }),
-            (InvalidAssets, "map", {
+            ("map", {
                 "tiles": [],
                 "tile_borders": [[2, 1]]
             }),
             # Impossible to achieve n_visible_combos=2:
-            (InvalidAssets, "races", []),
-            (InvalidAssets, "abilities", []),
+            ("races", []),
+            ("abilities", []),
         ]
-        for exc_type, key, value in invalid_fields:
+        for key, value in invalid_fields:
             invalid_assets = {**TINY_ASSETS, key: value}
-            with self.assertRaises(exc_type):
+            with self.assertRaises(ValidationError):
                 validate(invalid_assets)